requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.8
Django==4.0
emoji==1.6.1
fake-useragent==0.1.11
idna==3.3
lxml==4.6.4
mysql-connector-python==8.0.27
mysqlclient==2.1.0
protobuf==3.19.1
requests==2.26.0
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.7
